requirements.txt
fastapi[all]
uvicorn
sqlalchemy
alembic
gunicorn
python-dotenv
postgres
# psycopg2
psycopg2-binary
neo4j
tree-sitter<=0.21.3
tree-sitter-languages
tqdm
grep-ast
pygments
networkx
blar-graph
openai
duckduckgo-search
uuid6
aiohttp
langchain
langchain-openai
langchain-community
pytrends
wikipedia-api
langgraph
langchain-postgres
pymongo
firebase_admin
portkey_ai
gitPython
PyGithub
google-cloud-secret-manager
pre-commit
isort
black
ruff
pylint
bandit